Verschiebe `FilmlistenEintrag` nach filminfo.py
